remove duped tests, separate unit tests, etc
